LayoutTests/platform/wk2/imported/w3c/indexeddb/idbfactory_deleteDatabase3-expected.txt


FAIL IDBFactory.deleteDatabase() - success event assert_equals: oldVersion expected (number) 9 but got (undefined) undefined
